+++ integration_test/functions/src/v1/testLab-utils.ts
@@ -1,7 +1,5 @@
+import fetch from 'node-fetch';
 import * as admin from 'firebase-admin';
-import * as http from 'http';
-import * as https from 'https';
-import * as utils from './test-utils';
 
 interface AndroidDevice {
   androidModelId: string;
@@ -18,22 +16,31 @@ const TESTING_API_SERVICE_NAME = 'testing.googleapis.com';
  *
  * @param projectId Project for which the test run will be created
  * @param testId Test id which will be encoded in client info details
+ * @param accessToken accessToken to attach to requested for authentication
  */
-export async function startTestRun(projectId: string, testId: string) {
-  const accessToken = await admin.credential
-    .applicationDefault()
-    .getAccessToken();
+export async function startTestRun(
+  projectId: string,
+  testId: string,
+  accessToken: string
+) {
   const device = await fetchDefaultDevice(accessToken);
   return await createTestMatrix(accessToken, projectId, testId, device);
 }
 
-async function fetchDefaultDevice(
-  accessToken: admin.GoogleOAuthAccessToken
-): Promise<AndroidDevice> {
-  const response = await utils.makeRequest(
-    requestOptions(accessToken, 'GET', '/v1/testEnvironmentCatalog/ANDROID')
+async function fetchDefaultDevice(accessToken: string): Promise<AndroidDevice> {
+  const resp = await fetch(
+    `https://${TESTING_API_SERVICE_NAME}/v1/testEnvironmentCatalog/ANDROID`,
+    {
+      headers: {
+        Authorization: 'Bearer ' + accessToken,
+        'Content-Type': 'application/json',
+      },
+    }
   );
-  const data = JSON.parse(response);
+  if (!resp.ok) {
+    throw new Error(resp.statusText);
+  }
+  const data = await resp.json();
   const models = data?.androidDeviceCatalog?.models || [];
   const defaultModels = models.filter(
     (m) =>
@@ -58,17 +65,12 @@ async function fetchDefaultDevice(
   } as AndroidDevice;
 }
 
-function createTestMatrix(
-  accessToken: admin.GoogleOAuthAccessToken,
+async function createTestMatrix(
+  accessToken: string,
   projectId: string,
   testId: string,
   device: AndroidDevice
-): Promise<string> {
-  const options = requestOptions(
-    accessToken,
-    'POST',
-    '/v1/projects/' + projectId + '/testMatrices'
-  );
+): Promise<void> {
   const body = {
     projectId,
     testSpecification: {
@@ -96,21 +98,19 @@ function createTestMatrix(
       },
     },
   };
-  return utils.makeRequest(options, JSON.stringify(body));
-}
-
-function requestOptions(
-  accessToken: admin.GoogleOAuthAccessToken,
-  method: string,
-  path: string
-): https.RequestOptions {
-  return {
-    method,
-    hostname: TESTING_API_SERVICE_NAME,
-    path,
-    headers: {
-      Authorization: 'Bearer ' + accessToken.access_token,
-      'Content-Type': 'application/json',
-    },
-  };
+  const resp = await fetch(
+    `https://${TESTING_API_SERVICE_NAME}/v1/projects/${projectId}/testMatrices`,
+    {
+      method: 'POST',
+      headers: {
+        Authorization: 'Bearer ' + accessToken,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(body),
+    }
+  );
+  if (!resp.ok) {
+    throw new Error(resp.statusText);
+  }
+  return;
 }